Improvements to merge command (#567)
Bug to be fixed: after the merge command is run and the player is merged together, when that player splits again, running the merge command in that player will give "else console.log("[Console] Player " + client.name + " is no longer force merging");" and that player won't merge. So you'd need to run the merge command again for it to work.